+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Compatibility:
 * Fix `Regexp` issue which raised syntax error instead of `RegexpError` (#2066).
 * Handle `Object#autoload` when autoload itself (#1616, @ssnickolay)
 * Skip upgraded default gems while loading RubyGems (#2075).
+* Verify that gem paths are correct before loading RubyGems (#2075).
 
 Performance: